refactor(kafka): created new universal consumer factory
